DT-835 Unit test is better if we are mocking all beans
